requirements.txt
altgraph==0.17.4
certifi==2025.4.26
charset-normalizer==3.4.2
idna==3.10
macholib==1.16.3
modulegraph==0.19.6
packaging==25.0
py2app==0.28.8
pyobjc-core==11.1
pyobjc-framework-Cocoa==11.1
python-dotenv==1.1.0
requests==2.32.3
rumps==0.4.0
urllib3==2.4.0
